question_paper_analyzer/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
pydantic
python-dotenv
aiofiles
loguru
tenacity
